\name{Blink2}

\docType{data}

\alias{Blink2}

\title{Blink data (2nd example) from Preece (1982, Table 3)}

\description{
This dataset presents paired data corresponding to average blink-rate per minute of 12 subjects in an experiment of a visual motor task. They had to steer a pencil along a moving track. Each subject was tested under two conditions : a straight track and an oscillating one. 
Data about blink-rate during a pre-experimental resting are also available. Subjects 1 and 2 then appear less extreme than in the dataset Blink.
}

\usage{data(Blink2)}

\format{A dataframe with 12 rows and 4 columns: 
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab \cr
[,2] \tab Resting \tab numeric \tab blink rate in pre-experimental condition \cr
[,3] \tab Straight \tab numeric \tab blink rate in first condition \cr
[,4] \tab Oscillating \tab numeric \tab blink rate in second condition \cr
}
}

\seealso{Blink}

\source{
Drew, G.C. (1951) Variations in blink-rate during visual-motor tasks. Quarterly Journal of Experimental Psychology, 3, 73-88.
}

\references{
Preece D.A. (1982) t is for trouble (and textbooks): a critique of some examples of the paired-samples t-test. The Statistician, 31 (2), 169-195.
}


\keyword{datasets}
