\name{Datalcoholic}

\docType{data}

\alias{Datalcoholic}

\title{Datalcoholic: a dataset of paired datasets}

\description{
This dataset presents for teaching purposes 50 paired datasets available in different R packages.
}

\usage{data(Datalcoholic)}

\format{A dataframe with 4 columns.
\tabular{rll}{
[,1] \tab Dataset \tab name of the dataset \cr
[,2] \tab Package \tab name of the package \cr
[,3] \tab Topic \tab corresponding discipline (marketing, medicine...) \cr
[,4] \tab NumberPairs \tab size of the (paired) sample \cr
}
}

\examples{
data(Datalcoholic)
show(Datalcoholic)
}

\keyword{datasets}
