\name{GDO}

\docType{data}

\alias{GDO}

\title{Agreement study}

\description{
This dataset gives the same measurements in 3 days corresponding to a reproductibility study
for 18 subjects.
}

\usage{data(GDO)}

\format{A dataframe with 18 rows and 4 columns.
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab Anonymous subjects \cr
[,2] \tab Day1 \tab numeric \tab Measurement first day\cr
[,3] \tab Day2 \tab numeric \tab Measurement second day\cr
[,4] \tab Day3 \tab numeric \tab Measurement third day\cr
}
}

\source{Private communication. Samuel Rota, CRIS, Lyon 1 University, FRANCE}


\keyword{datasets}
