\name{HORSEbegin}

\docType{data}

\alias{HORSEbegin}

\title{Actual and imaginary performances in equitation}

\description{
This dataset gives the actual and motor imaginary performances (time) in equitation for 8 beginners.
}

\usage{data(HORSEbegin)}

\format{A dataframe with 8 rows and 3 columns.
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab Anonymous subjects \cr
[,2] \tab Actual \tab numeric \tab Actual performance (sec.)\cr
[,3] \tab Imaginary \tab numeric \tab Imaginary performance (sec.)\cr
}
}

\source{Private communication. Aymeric Guillot, CRIS, Lyon 1 University, FRANCE.}

\references{
M. Louis, C. Collet, S. Champely, and A. Guillot. Differences in motor imagery time when predicting task duration. 
Research Quarterly for Exercise and Sport, 2010.
}


\keyword{datasets}
