\name{lambda.table}

\docType{data}

\alias{lambda.table}

\title{Parameters for Generalised Lambda Distributions}

\description{
This dataset gives the parameters for specific 8 Generalized Tukey-lambda distributions with zero mean and unit variance
 useful for simulation studies as given in Bonett and Seier (2003).
}

\usage{data(lambda.table)}

\format{A dataframe with 8 rows (distributions) and 4 columns (parameters).

}

\references{Bonett, D.G. and Seier, E. (2003) Statistical inference for a ratio of dispersions using paired samples. Journal of Educational and Behavioral Statistics, 28, 21-30.}


\keyword{datasets}
