\name{PairedData-internal}
\title{Internal PairedData objects}

\alias{var1.test}
\alias{var2.test}


\alias{validPairedObject}

\alias{winvar.Z}

\alias{winval}

\alias{wincor}

\alias{winvar}

\alias{summaryhelper}

\alias{unroll}

\alias{unrollG}

\alias{yuen1.test}
\alias{yuen2.test}
\alias{yuenp.test}

\alias{paired.summary}

\alias{paired.effect.size}

\alias{paired.slidingchart}






\description{Internal PairedData objects.}
\details{These are not to be called by the user.}
\keyword{internal}
