\name{Sleep}

\docType{data}

\alias{Sleep}

\title{Sleep hours data from Preece (1982, Table 16)}

\description{
This dataset presents paired data corresponding to the sleep hours gained by 10 patients (these are differences indeed) using two isomers (Dextro- and Laevo-). These data from Student were studied by Fischer (1925). Read the paper of Preece (1982, section 9) for a complete understanding of this quite complex situation.
}

\usage{data(Sleep)}

\format{A dataframe with 10 rows and 2 columns: 
\tabular{rlll}{
[,1] \tab Dextro \tab numeric \tab sleep hour gain \cr
[,2] \tab Laevo \tab numeric \tab sleep hour gain \cr
}
}

\source{
Fisher, R.A. (1925) Statistical Metods for Research Workers. Oliver and Boyd: Edinburgh.
}

\references{
Preece, D.A. (1982) t is for trouble (and textbooks): a critique of some examples of the paired-samples t-test. The Statistician, 31 (2), 169-195.
}


\keyword{datasets}
