\name{anscombe2}

\docType{data}

\alias{anscombe2}

\title{Teaching the paired t test}

\description{
This dataset presents four sets of paired samples (n=15), giving the same t statistic (t=2.11) and thus the same
p-value whereas their situations are really different (differences in variances, clustering, heteroscedasticity). 
The importance of plotting data is thus stressed.
The name is given from the famous Anscombe's dataset created to study simple linear regression.
}

\usage{data(anscombe2)}

\format{A dataframe with 15 rows, 8 numeric columns and 1 column (factor) giving a label for the subjects : X1, X2, X3, X4, Y1, Y2, Y3, Y4, Subjects.
}

\source{S. Champely, CRIS, Lyon 1 University, FRANCE}

\references{F. Anscombe, Graphs in statistical analysis. The American Statistican, 27, 17-21.}


\keyword{datasets}
