\name{iceskating}

\docType{data}

\alias{iceskating}

\title{IceSkating speed study}

\description{
This dataset gives the speed measurement for 7 iceskating dancers using the return leg in flexion
or in extension.
}

\usage{data(iceskating)}

\format{A dataframe with 18 rows and 3 columns.
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab Anonymous subjects \cr
[,2] \tab Extension \tab numeric \tab Speed when return leg in extension (m/sec) \cr
[,3] \tab Flexion \tab numeric \tab Speed when return leg in flexion (m/sec) \cr
}
}

\source{Private communication. Karine Monteil, CRIS, Lyon 1 University, FRANCE}

\references{
Haguenauer, M., Legreneur, P., Colloud, F. and Monteil, K.M. Characterisation of
the Push-off in Ice Dancing: Influence of the Support Leg extension on Performance.
Journal of Human Movement Studies, 43, 197-210, 2002.
}


\keyword{datasets}
