\name{PairedData-internal}
\title{Internal PairedData objects}


\alias{validPairedObject}

\alias{winval}

\alias{wincor}

\alias{winvar}

\alias{unroll}

\alias{unrollG}

\alias{yuen1.test}
\alias{yuen2.test}
\alias{yuenp.test}

\alias{paired.summary}

\alias{paired.effect.size}

\alias{paired.plotBA}
\alias{paired.plotCor}
\alias{paired.plotMcNeil}
\alias{paired.plotProfiles}
\alias{paired.plotSliding}






\description{Internal PairedData objects.}
\details{These are not to be called by the user.}
\keyword{internal}
