\name{PairedData-package}
\alias{PairedData-package}
\alias{PairedData}
\docType{package}
\title{A package for visualising and analysing paired (numeric) data}
\description{
A package for visualising and analysing paired (numeric) data
}
\details{
\tabular{ll}{
Package: \tab PairedData\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2011-12-16\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Stephane Champely

Maintainer: Stephane Champely<champely@univ-lyon1.fr>

}

