\name{PairedData-package}
\alias{PairedData-package}
\alias{PairedData}
\docType{package}
\title{
\packageTitle{PairedData}
}
\description{
\packageDescription{PairedData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PairedData}
\packageIndices{PairedData}
}
\author{
\packageAuthor{PairedData}

Maintainer: \packageMaintainer{PairedData}
}
\keyword{ package }
