\name{PairedData-package}
\alias{PairedData-package}
\alias{PairedData}
\docType{package}
\title{A package for visualising and analysing paired data.
}

\description{
A package for visualising and analysing paired (mainly numeric) data with S4 class.

}

\details{
\tabular{ll}{
Package: \tab PairedData\cr
Type: \tab Package\cr
Version: \tab 0.9.9\cr
Date: \tab 2012-05-01\cr
License: \tab GPL \cr
Depends: \tab methods\cr
}
}


\author{
Stephane Champely

Maintainer: Stephane Champely<champely@univ-lyon1.fr>
}


\keyword{package}
