\name{PrisonStress}

\docType{data}

\alias{PrisonStress}

\title{Stress in prison}

\description{
This dataset gives the PSS (stress measurement) for 26 people in prison at the entry and at the exit. 
Part of these people were physically trained during their imprisonment.
}

\usage{data(PrisonStress)}

\format{A dataframe with 26 rows and 4 columns.
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab Anonymous subjects \cr
[,2] \tab Group \tab factor \tab Sport or Control\cr
[,3] \tab PSSbefore \tab numeric \tab Stress measurement before training\cr
[,4] \tab PSSafter \tab numeric \tab Stress measurement after training\cr
}
}

\source{Private communication. Charlotte Verdot, CRIS, Lyon 1 University, FRANCE}

\references{
Verdot, C., Champely, S., Massarelli, R. and Clement, M. (2008) Physical activities in prison as 
a tool to ameliorate detainees mood and well-being. 
International Review on Sport and Violence, 2.
}


\keyword{datasets}
