\name{rugby}

\docType{data}

\alias{rugby}

\title{Agreement study}

\description{
This dataset gives the ratings on a continuous ten-points scale of two experts about 93 actions
 during several rugby union matches.
}

\usage{data(rugby)}

\format{A dataframe with 93 rows and 3 columns.
\tabular{rlll}{
[,1] \tab EXPERT.1 \tab numeric \tab First expert ratings \cr
[,2] \tab EXPERT.2 \tab numeric \tab Second expert ratings \cr
[,3] \tab Actions \tab factor \tab Subject label\cr
}
}

\source{Private communication. Mickael Campo, CRIS, Lyon 1 University, FRANCE.}


\keyword{datasets}
