\name{shoulder}

\docType{data}

\alias{shoulder}

\title{Shoulder flexibility}

\description{
This dataset gives the flexibility for the right and left shoulders in 15 swimmers and
15 sedentary people.
}

\usage{data(shoulder)}

\format{A dataframe with 12 rows and 4 columns.
\tabular{rlll}{
[,1] \tab Subject \tab factor \tab Anonymous subjects \cr
[,2] \tab Group \tab factor \tab Swimmer or Control\cr
[,3] \tab Right \tab numeric \tab Right shoulder flexibility (deg.)\cr
[,4] \tab Left \tab numeric \tab Left shoulder flexibility (deg.)\cr
}
}

\source{Private communication. Karine Monteil, CRIS, Lyon 1 University, FRANCE}

\references{Monteil, K., Taiar, R., Champely, S. and Martin, J. (2002) Competitive swimmers versus sedentary people: 
a predictive model based upon normal shoulders flexibility. Journal of Human Movement Studies, 43 , 17-34.
}


\keyword{datasets}
